[ALPS09440596] cam:HP9 porting update
[Description]
0. fix mipi timing
1. enable AEB mode & update frame desc
2. fix AEB & stagger mode FL related param.
3. update cus4 (using FCM_8 without AEB)
4. fix cus10 bayer order (3M mode with crop)
5. fix DCG binning ratio
6. for linearity, fix min. & max. & step by mode
7. exposure max is bounded at currently patch
   (long exposure is disable)
   (which need fll_lshift patch:10171881)
8. rm ana_gain_max in mode struct (not used)
9. modify AEB VCDT order in framedesc
10. for AEB mode, read_margin u8->u32

[Test]
UT

MTK-Commit-Id: 847205d061d11ceec8cd4f2ad85daa3800b0fd2e

CR-Id: ALPS09440596
Feature: [Android Default] Camera Application Basic Functions

+++ imgsensor/src-isp8s/adaptor-fw-types.h
@@ -399,7 +399,7 @@ struct fw_mode_info {
 	u16 max_framerate;
 	u32 mipi_pixel_rate;
 	u32 readout_length;
-	u8 read_margin;
+	u32 read_margin;
 	u32 framelength_step;
 	u32 coarse_integ_step;
 	u32 min_exposure_line;
