+++ imgsensor/src-isp8s/adaptor-subdrv.h
@@ -324,7 +324,7 @@ struct subdrv_mode_struct {
 	u16 max_framerate;
 	u32 mipi_pixel_rate;
 	u32 readout_length;
-	u8 read_margin;
+	u32 read_margin;
 	u32 framelength_step;
 	u32 coarse_integ_step;
 	u32 min_exposure_line;
